- br-installconf.sh: added option (-L) to update linux configuration
   The linux config is assembled from snippets under site/configs.
   Then we remove .stamp_configured from the linux directory so
   that the kernel is reconfigured.

+++ scripts/br-installconf.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-TEMP=`getopt -o a:hpfdO -n "$0" -- "$@"`
+TEMP=`getopt -o a:hpfdOL -n "$0" -- "$@"`
 
 if [ $? != 0 ] ; then
 	echo "getopt error - terminating..." >&2;
@@ -15,6 +15,7 @@ while true; do
 			echo "          -a zynq | x86_64 | i686"
 			echo "          -p omit patching buildroot (if already done)"
 			echo "          -f force rerun (if already done)"
+			echo "          -L just update linux configuration"
 			echo "          -d dry-run"
 			echo "          -O out-of-tree build (not too useful, though)"
 			exit 0
@@ -35,10 +36,6 @@ while true; do
 			case "$2" in
 				"zynq" | "i686" | "x86_64")
 					ARCH="$2"
-					case "$2" in
-						"zynq")            KARCH="$2"  ;;
-						"i686" | "x86_64") KARCH="x86" ;;
-					esac
 				;;
 				*)
 					echo "Unsupported arch/config '$2'" ;
@@ -48,7 +45,11 @@ while true; do
 			shift 2
 		;;
 		-O )
-			OOF_TREE=y
+			OOF_TREE="y"
+			shift
+		;;
+		-L )
+			LCONF="y"
 			shift
 		;;
 		--) shift; break
@@ -58,11 +59,68 @@ while true; do
 	esac
 done
 
+get_makevar() {
+	make ${MKOUTDIR} -f - print-var <<EOF
+include Makefile
+print-var:
+	@echo \$(${1})
+EOF
+}
+
+restore() {
+	rm ${CONF_DIR}/.config ${CONF_DIR}/.config.orig
+	if [ -f ${CONF_DIR}/.config.bup ] ; then
+		mv ${CONF_DIR}/.config.bup ${CONF_DIR}/.config
+	fi
+	if [ -f ${CONF_DIR}/.config.old.bup ] ; then
+		mv ${CONF_DIR}/.config.old.bup ${CONF_DIR}/.config.old
+	fi
+	if [ -f ${CONF_DIR}/linux-${LINUX_VER}.config.bup ] ; then
+		mv ${CONF_DIR}/linux-${LINUX_VER}.config.bup ${CONF_DIR}/linux-${LINUX_VER}.config
+	fi
+}
+
+install_linux_conf() {
+	if [ -f ${CONF_DIR}/linux-${LINUX_VER}.config ] ; then
+		mv ${CONF_DIR}/linux-${LINUX_VER}.config ${CONF_DIR}/linux-${LINUX_VER}.config.bup
+	fi
+	if [ -f site/config/linux-${LINUX_VER}-common.config -a -f site/config/linux-${LINUX_VER}-${KARCH}.config ] ; then
+		cat site/config/linux-${LINUX_VER}-common.config site/config/linux-${LINUX_VER}-${KARCH}.config  > ${CONF_DIR}/linux-${LINUX_VER}.config
+		cp ${CONF_DIR}/linux-${LINUX_VER}.config ${CONF_DIR}/linux-${LINUX_VER}.config.orig
+		if [ -f ${CONF_DIR}/output/build/linux-${LINUX_VER}/.config ] ; then
+			cp -b ${CONF_DIR}/output/build/linux-${LINUX_VER}/.config ${CONF_DIR}/output/build/linux-${LINUX_VER}/.config.bup
+			rm -f ${CONF_DIR}/output/build/linux-${LINUX_VER}/.stamp_configured
+		fi
+		rm -f 
+	else
+		echo "Error: linux config snippets for linux-${LINUX_VER} not found" >&2
+		if [ -z "${LCONF}" ] ; then
+			restore;
+		fi
+		exit 1
+	fi
+}
+
+if [ -z "${OOF_TREE}" -a -f ./.stamp_br_installconf ] ; then
+	BR_ARCH_VARIANT=`cat ./.stamp_br_installconf`
+	if [ -z "${ARCH}" ] ; then
+		ARCH="${BR_ARCH_VARIANT}"
+	else if [ -n "${BR_ARCH_VARIANT}" -a "${ARCH}" != "${BR_ARCH_VARIANT}" ] ; then
+			echo "Error: arch '${ARCH}' requested but already configured for '${BR_ARCH_VARIANT}'"
+			exit 1;
+		fi
+	fi
+fi
 if [ -z "$ARCH" ] ; then
 	echo "Error: No target architecture given" >&2
 	echo "Usage: $0 -a <zynq | i686 | x86_64>" >&2
 	exit 1;
 fi
+
+case "${ARCH}" in
+	"i686" | "x86_64") KARCH="x86"      ;;
+	*)                 KARCH="${ARCH}"  ;;
+esac
 
 ##NOTE the buildroot out-of-tree feature is not
 ##     really useful since it still duplicates
@@ -79,25 +137,28 @@ else
 	CONF_DIR=.
 fi
 
-if [ -f ${CONF_DIR}/.stamp_br_installconf ] ; then
-	if [ -z "$FORCE" ] ; then
-		echo "Error: $0 was already executed (use -f to force, -fp to avoid repatching)!" >&2
-		exit 1;
-	fi
-	rm ${CONF_DIR}/.stamp_br_installconf
-	if [ -z "$NOPATCH" ] ; then
-		rm .stamp_br_patched
-	fi
-fi
+if [ "$LCONF"x != "yx" ] ; then
 
-if [ -f .stamp_br_patched -o "$NOPATCH" = "y" ] ; then
-	if [ "$NOPATCH" = "y" ] ; then
-		touch .stamp_br_patched
+	if [ -f ${CONF_DIR}/.stamp_br_installconf ] ; then
+		if [ -z "$FORCE" ] ; then
+			echo "Error: $0 was already executed (use -f to force, -fp to avoid repatching)!" >&2
+			exit 1;
+		fi
+		rm ${CONF_DIR}/.stamp_br_installconf
+		if [ -z "$NOPATCH" ] ; then
+			rm .stamp_br_patched
+		fi
 	fi
-else
-	if cat site/br-patches/buildroot* | patch ${DRY_RUN} -p0 -b ; then
-		if [ -z "${DRY_RUN}" ] ; then
+
+	if [ -f .stamp_br_patched -o "$NOPATCH" = "y" ] ; then
+		if [ "$NOPATCH" = "y" ] ; then
 			touch .stamp_br_patched
+		fi
+	else
+		if cat site/br-patches/buildroot* | patch ${DRY_RUN} -p0 -b ; then
+			if [ -z "${DRY_RUN}" ] ; then
+				touch .stamp_br_patched
+			fi
 		fi
 	fi
 fi
@@ -108,41 +169,28 @@ if [ $? != 0 ]; then
 	exit 1
 fi
 
-restore() {
-	rm ${CONF_DIR}/.config ${CONF_DIR}/.config.orig
-	if [ -f ${CONF_DIR}/.config.bup ] ; then
-		mv ${CONF_DIR}/.config.bup ${CONF_DIR}/.config
-	fi
-	if [ -f ${CONF_DIR}/.config.old.bup ] ; then
-		mv ${CONF_DIR}/.config.old.bup ${CONF_DIR}/.config.old
-	fi
-	if [ -f ${CONF_DIR}/linux-${LINUX_VER}.config.bup ] ; then
-		mv ${CONF_DIR}/linux-${LINUX_VER}.config.bup ${CONF_DIR}/linux-${LINUX_VER}.config
-	fi
-}
+if [ "$LCONF"x != "yx" ] ; then
 
-if [ -f ${CONF_DIR}/.config ] ; then
-	mv ${CONF_DIR}/.config ${CONF_DIR}/.config.bup
-fi
-cat site/config/br-${BR_VER}-${ARCH}.config site/config/br-${BR_VER}-common.config > ${CONF_DIR}/.config
-if [ $? != 0 ] ; then
-	echo "Error: unable to install .config file" >&2
-	restore
-	exit 1
-fi
-cp ${CONF_DIR}/.config ${CONF_DIR}/.config.orig
-if [ -f ${CONF_DIR}/.config.old ] ; then
-	cp ${CONF_DIR}/.config.old ${CONF_DIR}/.config.old.bup
+	if [ -f ${CONF_DIR}/.config ] ; then
+		mv ${CONF_DIR}/.config ${CONF_DIR}/.config.bup
+	fi
+	cat site/config/br-${BR_VER}-${ARCH}.config site/config/br-${BR_VER}-common.config > ${CONF_DIR}/.config
+	if [ $? != 0 ] ; then
+		echo "Error: unable to install .config file" >&2
+		restore
+		exit 1
+	fi
+	cp ${CONF_DIR}/.config ${CONF_DIR}/.config.orig
+	if [ -f ${CONF_DIR}/.config.old ] ; then
+		cp ${CONF_DIR}/.config.old ${CONF_DIR}/.config.old.bup
+	fi
+
+	make ${MKOUTDIR} olddefconfig
+
 fi
 
-make ${MKOUTDIR} olddefconfig
+LINUX_VER=`get_makevar LINUX_VERSION`
 
-LINUX_VER=`make ${MKOUTDIR} -f - print-linux-version <<"EOF"
-include Makefile
-print-linux-version:
-	@echo $(LINUX_VERSION)
-EOF
-`
 if [ $? != 0 -o -z "${LINUX_VER}" ]; then
 	echo "Error: unable to determine linux version" >&2
 	restore;
@@ -160,25 +208,18 @@ fi
 
 echo "BR version $BR_VER"
 echo "LI version '$LINUX_VER'"
+echo "ARCH ${ARCH}"
+echo "KARCH ${KARCH}"
 
 if [ -n "${DRY_RUN}" ] ; then
 	restore;
 	exit 0
 fi
 
+install_linux_conf;
 
-if [ -f ${CONF_DIR}/linux-${LINUX_VER}.config ] ; then
-	mv ${CONF_DIR}/linux-${LINUX_VER}.config ${CONF_DIR}/linux-${LINUX_VER}.config.bup
+if [ "$LCONF"x != "yx" ] ; then
+	echo "${ARCH}" > ${CONF_DIR}/.stamp_br_installconf
+
+	echo "Now type 'make${SP}${MKOUTDIR}' to build"
 fi
-if [ -f site/config/linux-${LINUX_VER}-common.config -a -f site/config/linux-${LINUX_VER}-${KARCH}.config ] ; then
-	cat site/config/linux-${LINUX_VER}-common.config site/config/linux-${LINUX_VER}-${KARCH}.config  > ${CONF_DIR}/linux-${LINUX_VER}.config
-	cp ${CONF_DIR}/linux-${LINUX_VER}.config ${CONF_DIR}/linux-${LINUX_VER}.config.orig
-else
-	echo "Error: linux config snippets for linux-${LINUX_VER} not found" >&2
-	restore;
-	exit 1
-fi
-
-touch ${CONF_DIR}/.stamp_br_installconf
-
-echo "Now type 'make${SP}${MKOUTDIR}' to build"
